Move remaining archive subcommands to forest-tool
